backend/requirements.txt
flask
flask-cors
flask-socketio
PyMySQL
python-dotenv
PyJWT
gunicorn
flask-limiter
werkzeug
